requirements.txt
python-dotenv>=1.0
httpx>=0.27,<0.28
beautifulsoup4>=4.12
yfinance>=0.2
pandas>=2.2
numpy>=1.26
pytest>=8.0
strands-agents==1.9.1
openai==1.40.0
pydantic>=2.0
